+++ AINDY/routes/__init__.py
@@ -20,6 +20,7 @@ from routes.auth_router import router as auth_router
 from routes.masterplan_router import router as masterplan_router
 from routes.memory_router import router as memory_router
 from routes.memory_metrics_router import router as memory_metrics_router
+from routes.memory_trace_router import router as memory_trace_router
 from routes.identity_router import router as identity_router
 
 
@@ -46,6 +47,7 @@ ROUTERS = [
     masterplan_router,
     memory_router,
     memory_metrics_router,
+    memory_trace_router,
     identity_router,
 ]